+++ gradio_tools/__init__.py
@@ -1,1 +1,16 @@
-from gradio_tools.tools import GradioTool, StableDiffusionTool, ImageCaptioningTool, ImageToMusicTool, StableDiffusionPromptGeneratorTool, WhisperTool, ClipInterrogatorTool, TextToVideoTool
+from gradio_tools.tools import (ClipInterrogatorTool, GradioTool,
+                                ImageCaptioningTool, ImageToMusicTool,
+                                StableDiffusionPromptGeneratorTool,
+                                StableDiffusionTool, TextToVideoTool,
+                                WhisperAudioTranscriptionTool)
+
+__all__ = [
+    "GradioTool",
+    "StableDiffusionTool",
+    "ClipInterrogatorTool",
+    "ImageCaptioningTool",
+    "ImageToMusicTool",
+    "WhisperAudioTranscriptionTool",
+    "StableDiffusionPromptGeneratorTool",
+    "TextToVideoTool",
+]
